feat: implement body overflow control on tile expand/collapse and add loader types
