+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from "./components"
 export * from "./core"
 export * from "./services"
 export * from "./SdkSwiper"
+export * from "./loader"
